backend/requirements.txt
django
GDAL==3.4.1
psycopg2
djangorestframework
djangorestframework-gis
django-leaflet
django-rest-swagger
